pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "minecraft-ops-mcp"
version = "0.3.0"
description = "An MCP server for Minecraft operations via MCSManager, RCON, and MSMP."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "minecraft-ops-mcp contributors" }]
keywords = ["mcp", "minecraft", "mcsmanager", "rcon", "msmp", "operations"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: System Administrators",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: System :: Systems Administration",
  "Typing :: Typed",
]
dependencies = [
  "mcp==1.27.0",
  "websocket-client==1.9.0",
]

[project.scripts]
minecraft-ops-mcp = "minecraft_ops_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
minecraft_ops_mcp = ["py.typed"]
